% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLaunch.R
\docType{package}
\name{shinyLaunch}
\alias{shinyLaunch}
\alias{shinyLaunch-package}
\title{shiny dashboard launch}
\usage{
shinyLaunch()
}
\description{
default shiny launcher
}
\details{
EmptyPackage

Default shiny app launcher
}
